+++ ticker/src/main/java/com/robinhood/ticker/TickerColumnManager.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (C) 2016 Robinhood Markets, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,8 +20,8 @@ import android.graphics.Canvas;
 import android.graphics.Paint;
 
 import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * In ticker, each character in the rendered text is represented by a {@link TickerColumn}. The
@@ -32,70 +32,87 @@ import java.util.Map;
  *
  * @author Jin Cao, Robinhood
  */
+@SuppressWarnings("ForLoopReplaceableByForEach")
 class TickerColumnManager {
     final ArrayList<TickerColumn> tickerColumns = new ArrayList<>();
     private final TickerDrawMetrics metrics;
 
-    // The character list that dictates how to transition from one character to another.
-    private char[] characterList;
-    // A minor optimization so that we can cache the indices of each character.
-    private Map<Character, Integer> characterIndicesMap;
+    private TickerCharacterList[] characterLists;
+    private Set<Character> supportedCharacters;
 
     TickerColumnManager(TickerDrawMetrics metrics) {
         this.metrics = metrics;
     }
 
     /**
-     * @see {@link TickerView#setCharacterList(char[])}.
+     * @inheritDoc TickerView#setCharacterLists
      */
-    void setCharacterList(char[] characterList) {
-        this.characterList = characterList;
-        this.characterIndicesMap = new HashMap<>(characterList.length);
+    void setCharacterLists(String... characterLists) {
+        this.characterLists = new TickerCharacterList[characterLists.length];
+        for (int i = 0; i < characterLists.length; i++) {
+            this.characterLists[i] = new TickerCharacterList(characterLists[i]);
+        }
 
-        for (int i = 0; i < characterList.length; i++) {
-            characterIndicesMap.put(characterList[i], i);
+        this.supportedCharacters = new HashSet<>();
+        for (int i = 0; i < characterLists.length; i++) {
+            this.supportedCharacters.addAll(this.characterLists[i].getSupportedCharacters());
         }
     }
 
-    /**
-     * @return whether or not {@param text} should be debounced because it's the same as the
-     *         current target text of this column manager.
-     */
-    boolean shouldDebounceText(char[] text) {
-        final int newTextSize = text.length;
-        if (newTextSize == tickerColumns.size()) {
-            for (int i = 0; i < newTextSize; i++) {
-                if (text[i] != tickerColumns.get(i).getTargetChar()) {
-                    return false;
-                }
-            }
-            return true;
-        }
-        return false;
+    TickerCharacterList[] getCharacterLists() {
+        return characterLists;
     }
 
     /**
      * Tell the column manager the new target text that it should display.
      */
-    boolean setText(char[] text, boolean animate) {
-        if (characterList == null) {
-            throw new IllegalStateException("Need to call setCharacterList(char[]) first.");
+    void setText(char[] text) {
+        if (characterLists == null) {
+            throw new IllegalStateException("Need to call #setCharacterLists first.");
         }
 
-        final int newTextSize = text.length;
-        if (animate) {
-            // If we are animating, we don't want to remove old columns yet.
-            insertColumnsUpTo(newTextSize);
-        } else {
-            ensureColumnSize(newTextSize);
+        // First remove any zero-width columns
+        for (int i = 0; i < tickerColumns.size(); ) {
+            final TickerColumn tickerColumn = tickerColumns.get(i);
+            if (tickerColumn.getCurrentWidth() > 0) {
+                i++;
+            } else {
+                tickerColumns.remove(i);
+            }
         }
 
-        final int columnsSize = tickerColumns.size();
-        for (int i = 0; i < columnsSize; i++) {
-            tickerColumns.get(i).setTargetChar(i < newTextSize ? text[i] : TickerUtils.EMPTY_CHAR);
+        // Use Levenshtein distance algorithm to figure out how to manipulate the columns
+        final int[] actions = LevenshteinUtils.computeColumnActions(
+                getCurrentText(), text, supportedCharacters
+        );
+        int columnIndex = 0;
+        int textIndex = 0;
+        for (int i = 0; i < actions.length; i++) {
+            switch (actions[i]) {
+                case LevenshteinUtils.ACTION_INSERT:
+                    tickerColumns.add(columnIndex,
+                            new TickerColumn(characterLists, metrics));
+                    // Intentional fallthrough
+                case LevenshteinUtils.ACTION_SAME:
+                    tickerColumns.get(columnIndex).setTargetChar(text[textIndex]);
+                    columnIndex++;
+                    textIndex++;
+                    break;
+                case LevenshteinUtils.ACTION_DELETE:
+                    tickerColumns.get(columnIndex).setTargetChar(TickerUtils.EMPTY_CHAR);
+                    columnIndex++;
+                    break;
+                default:
+                    throw new IllegalArgumentException("Unknown action: " + actions[i]);
+            }
         }
+    }
 
-        return true;
+    void onAnimationEnd() {
+        for (int i = 0, size = tickerColumns.size(); i < size; i++) {
+            final TickerColumn column = tickerColumns.get(i);
+            column.onAnimationEnd();
+        }
     }
 
     void setAnimationProgress(float animationProgress) {
@@ -121,6 +138,15 @@ class TickerColumnManager {
         return width;
     }
 
+    char[] getCurrentText() {
+        final int size = tickerColumns.size();
+        final char[] currentText = new char[size];
+        for (int i = 0; i < size; i++) {
+            currentText[i] = tickerColumns.get(i).getCurrentChar();
+        }
+        return currentText;
+    }
+
     /**
      * This method will draw onto the canvas the appropriate UI state of each column dictated
      * by {@param animationProgress}. As a side effect, this method will also translate the canvas
@@ -131,33 +157,6 @@ class TickerColumnManager {
             final TickerColumn column = tickerColumns.get(i);
             column.draw(canvas, textPaint);
             canvas.translate(column.getCurrentWidth(), 0f);
-        }
-    }
-
-    /**
-     * Ensure that the number of columns matches {@param targetSize}.
-     */
-    void ensureColumnSize(int targetSize) {
-        final int columnSize = tickerColumns.size();
-        if (targetSize > columnSize) {
-            insertColumnsUpTo(targetSize);
-        } else {
-            for (int i = 0; i < columnSize - targetSize; i++) {
-                tickerColumns.remove(tickerColumns.size() - 1);
-            }
-        }
-    }
-
-    /**
-     * Insert (if applicable) columns until the number of columns match {@param targetSize}.
-     */
-    void insertColumnsUpTo(int targetSize) {
-        final int currentSize = tickerColumns.size();
-        if (targetSize > currentSize) {
-            final int toInsert = targetSize - currentSize;
-            for (int i = 0; i < toInsert; i++) {
-                tickerColumns.add(new TickerColumn(characterList, characterIndicesMap, metrics));
-            }
         }
     }
 }